TileModelCompiler: fix compile warnings

+++ src/osgEarthDrivers/engine_quadtree/TileModelCompiler.cpp
@@ -1352,9 +1352,9 @@ namespace
                 boundaryElevations.reserve( 2 * d.numRows );
 
                 //Compute the verts for the west side
-                for (int j = 0; j < d.numRows; j++)
+                for (int j = 0; j < (int)d.numRows; j++)
                 {
-                    for (int i = d.numCols-2; i <= d.numCols-1; i++)
+                    for (int i = (int)d.numCols-2; i <= (int)d.numCols-1; i++)
                     {                          
                         osg::Vec3d ndc( (double)(i - static_cast<int>(d.numCols-1))/(double)(d.numCols-1), ((double)j)/(double)(d.numRows-1), 0.0);                                                                        
 
@@ -1371,7 +1371,7 @@ namespace
                 }   
 
                 //The boundary verts are now populated, so go through and triangulate them add add the normals to the existing normal array
-                for (int j = 0; j < d.numRows-1; j++)
+                for (int j = 0; j < (int)d.numRows-1; j++)
                 {                    
                     int i00;
                     int i01;
@@ -1437,7 +1437,7 @@ namespace
                 boundaryElevations.reserve( 2 * d.numRows );
 
                 //Compute the verts for the east side
-                for (int j = 0; j < d.numRows; j++)
+                for (int j = 0; j < (int)d.numRows; j++)
                 {
                     for (int i = 0; i <= 1; i++)
                     {                           
@@ -1456,7 +1456,7 @@ namespace
                 }   
 
                 //The boundary verts are now populated, so go through and triangulate them add add the normals to the existing normal array
-                for (int j = 0; j < d.numRows-1; j++)
+                for (int j = 0; j < (int)d.numRows-1; j++)
                 {                    
                     int i00;
                     int i01;
@@ -1521,7 +1521,7 @@ namespace
                 //Compute the verts for the north side               
                 for (int j = 0; j <= 1; j++)
                 {
-                    for (int i = 0; i < d.numCols; i++)                    
+                    for (int i = 0; i < (int)d.numCols; i++)                    
                     {                           
                         osg::Vec3d ndc( (double)(i)/(double)(d.numCols-1), (double)(d.numRows -1 + j)/(double)(d.numRows-1), 0.0);
                         //osg::Vec3d ndc( (double)(i)/(double)(d.numCols-1), (double)(-static_cast<int>(j))/(double)(d.numRows-1), 0.0);                        
@@ -1539,7 +1539,7 @@ namespace
                 }   
 
                 //The boundary verts are now populated, so go through and triangulate them add add the normals to the existing normal array                
-                for (int i = 0; i < d.numCols-1; i++)
+                for (int i = 0; i < (int)d.numCols-1; i++)
                 {                    
                     int i00;                    
                     int j = d.numRows-1;
@@ -1602,9 +1602,9 @@ namespace
                 boundaryElevations.reserve( 2 * d.numCols );
 
                 //Compute the verts for the south side               
-                for (int j = d.numRows-2; j <= d.numRows-1; j++)
+                for (int j = (int)d.numRows-2; j <= (int)d.numRows-1; j++)
                 {
-                    for (int i = 0; i < d.numCols; i++)                    
+                    for (int i = 0; i < (int)d.numCols; i++)                    
                     {                           
                         osg::Vec3d ndc( (double)(i)/(double)(d.numCols-1), (double)(j - static_cast<int>(d.numRows-1))/(double)(d.numRows-1), 0.0);                                                
                         
@@ -1621,7 +1621,7 @@ namespace
                 }   
 
                 //The boundary verts are now populated, so go through and triangulate them add add the normals to the existing normal array                
-                for (int i = 0; i < d.numCols-1; i++)
+                for (int i = 0; i < (int)d.numCols-1; i++)
                 {                    
                     int i00;                    
                     int j = 0;
